server/requirements.txt
mcp[cli]>=1.0.0
python-dotenv>=1.0.0
uvicorn>=0.30.0
